+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/Krypton Custom Theme Generator/Controls Visuals/VisualCustomThemeBuilderForm.cs
@@ -87,6 +87,7 @@ internal partial class VisualCustomThemeBuilderForm : KryptonForm
         kcmbFlyout.Items.Add(KryptonScreenColorPicker.GetFlyoutStyleDisplayName(KryptonScreenColorPickerFlyoutStyle.Classic));
         kcmbFlyout.SelectedIndex = KryptonScreenColorPicker.DefaultFlyoutStyle == KryptonScreenColorPickerFlyoutStyle.Classic ? 1 : 0;
         knudMagnifierSize.Value = KryptonScreenColorPicker.DefaultMagnifierSize;
+        KryptonScreenColorPicker.BindColorFormatList(kclbColorFormats);
         _suppressPreview = false;
     }